Fix imports, move fmtVal (formerly scope/util.go) to hantek6022be as the only user at this point.
