+++ usb/hantek6022be/calibration.go
@@ -18,7 +18,6 @@ import (
 	"fmt"
 
 	"github.com/pkg/errors"
-	"github.com/zagrodzki/goscope/scope"
 )
 
 type calData struct {
