+++ usb/hantek6022be/util_test.go
@@ -12,7 +12,7 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
-package scope
+package hantek6022be
 
 import "testing"
